requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv
email-validator
python-multipart
pydantic[email]
alembic
asyncpg